test: split proposal card workflows
