--- tests/Feature/Chat/ProposalCardComponentTest.php
@@ -1,1452 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-use App\Enums\CustomFieldType;
-use App\Features\OnboardSeed;
-use App\Models\Company;
-use App\Models\CustomField;
-use App\Models\People;
-use App\Models\Task;
-use App\Models\User;
-use Filament\Facades\Filament;
-use Filament\Forms\Components\Field;
-use Illuminate\Support\Facades\Bus;
-use Illuminate\Support\Facades\DB;
-use Laravel\Ai\Tools\Request;
-use Laravel\Pennant\Feature;
-use Livewire\Livewire;
-use Relaticle\Chat\Enums\PendingActionOperation;
-use Relaticle\Chat\Enums\PendingActionStatus;
-use Relaticle\Chat\Livewire\Chat\ProposalCard;
-use Relaticle\Chat\Models\PendingAction;
-use Relaticle\Chat\Support\PlanReference;
-use Relaticle\Chat\Tools\Company\CreateCompanyTool;
-use Relaticle\Chat\Tools\People\CreatePersonTool;
-use Relaticle\Chat\Tools\Task\CreateTaskTool;
-use Relaticle\CustomFields\Data\CustomFieldSettingsData;
-use Relaticle\CustomFields\Data\VisibilityConditionData;
-use Relaticle\CustomFields\Data\VisibilityData;
-use Relaticle\CustomFields\Enums\ConditionSource;
-use Relaticle\CustomFields\Enums\VisibilityMode;
-use Relaticle\CustomFields\Enums\VisibilityOperator;
-use Spatie\LaravelData\DataCollection;
-
-mutates(ProposalCard::class);
-
-beforeEach(function (): void {
-    Feature::define(OnboardSeed::class, false);
-    $this->user = User::factory()->withPersonalTeam()->create();
-    $this->team = $this->user->currentTeam;
-    $this->actingAs($this->user);
-    Filament::setTenant($this->team);
-});
-
-/**
- * @param  array<string, mixed>  $action
- * @param  array<string, mixed>  $display
- */
-function proposalCardPa(User $user, array $action, array $display): PendingAction
-{
-    return PendingAction::query()->create([
-        'team_id' => $user->currentTeam->getKey(),
-        'user_id' => $user->getKey(),
-        'conversation_id' => null,
-        'action_class' => 'App\\Actions\\Company\\CreateCompany',
-        'operation' => PendingActionOperation::Create,
-        'entity_type' => 'company',
-        'action_data' => $action,
-        'display_data' => $display,
-        'status' => PendingActionStatus::Pending,
-        'expires_at' => now()->addMinutes(15),
-    ]);
-}
-
-/**
- * @param  list<string>  $names
- */
-function makeBatchCompanyProposal(User $user, array $names): PendingAction
-{
-    $records = array_map(static fn (string $name): array => ['name' => $name], $names);
-    $items = array_map(static fn (string $name): array => [
-        'title' => $name,
-        'summary' => "Create company \"{$name}\"",
-        'fields' => [['label' => 'Name', 'value' => $name]],
-    ], $names);
-
-    return proposalCardPa(
-        $user,
-        ['_batch' => true, 'records' => $records],
-        ['title' => 'Create Companies', 'summary' => 'Create '.count($names).' companies', 'items' => $items],
-    );
-}
-
-/**
- * @param  array<string, mixed>  $actionData
- */
-function makeTaskProposal(User $user, array $actionData): PendingAction
-{
-    return PendingAction::query()->create([
-        'team_id' => $user->currentTeam->getKey(),
-        'user_id' => $user->getKey(),
-        'conversation_id' => null,
-        'action_class' => 'App\\Actions\\Task\\CreateTask',
-        'operation' => PendingActionOperation::Create,
-        'entity_type' => 'task',
-        'action_data' => $actionData,
-        'display_data' => ['title' => 'Create Task', 'summary' => 'Create task', 'fields' => []],
-        'status' => PendingActionStatus::Pending,
-        'expires_at' => now()->addMinutes(15),
-    ]);
-}
-
-/**
- * Batch variant of makeTaskProposal: an `_batch` action_data carrying one
- * `records` entry per task (each may include a `custom_fields` map) with a
- * matching `display_data['items']` list, mirroring makeBatchCompanyProposal.
- *
- * @param  list<array<string, mixed>>  $records
- */
-function makeBatchTaskProposal(User $user, array $records): PendingAction
-{
-    $items = array_map(static function (array $record): array {
-        $title = (string) ($record['title'] ?? '');
-
-        return [
-            'title' => 'Create Task',
-            'summary' => "Create task \"{$title}\"",
-            'fields' => [['label' => 'Title', 'code' => 'title', 'value' => $title]],
-        ];
-    }, $records);
-
-    return PendingAction::query()->create([
-        'team_id' => $user->currentTeam->getKey(),
-        'user_id' => $user->getKey(),
-        'conversation_id' => null,
-        'action_class' => 'App\\Actions\\Task\\CreateTask',
-        'operation' => PendingActionOperation::Create,
-        'entity_type' => 'task',
-        'action_data' => ['_batch' => true, 'records' => $records],
-        'display_data' => ['title' => 'Create Tasks', 'summary' => 'Create '.count($records).' tasks', 'items' => $items],
-        'status' => PendingActionStatus::Pending,
-        'expires_at' => now()->addMinutes(15),
-    ]);
-}
-
-/**
- * The seeded task `status` field is a SINGLE_CHOICE with To do / In progress /
- * Done options, created for every team by the CreateTeamCustomFields listener.
- *
- * @return array{0: CustomField, 1: list<string>}
- */
-function seedTaskSingleChoiceField(mixed $team): array
-{
-    $status = CustomField::query()
-        ->where('tenant_id', $team->getKey())
-        ->where('entity_type', 'task')
-        ->where('code', 'status')
-        ->with('options')
-        ->first();
-
-    expect($status)->not->toBeNull('seeded task status field is required for this test');
-
-    $optionIds = $status->options->map(fn (mixed $option): string => (string) $option->id)->values()->all();
-
-    expect($optionIds)->not->toBeEmpty();
-
-    return [$status, $optionIds];
-}
-
-/**
- * Build a task custom field that is only visible when the seeded `status`
- * field equals "Done" — a cross-field (sibling) visibility condition. Under
- * `->only([$code])` the sibling is absent from the scoped form, so the
- * condition must fail open rather than throw.
- */
-function seedTaskFieldWithVisibilityCondition(mixed $team): CustomField
-{
-    [$status] = seedTaskSingleChoiceField($team);
-
-    return CustomField::query()->create([
-        'tenant_id' => $team->getKey(),
-        'entity_type' => 'task',
-        'code' => 'completion_note',
-        'name' => 'Completion note',
-        'type' => 'text',
-        'sort_order' => 99,
-        'validation_rules' => [],
-        'active' => true,
-        'system_defined' => false,
-        'settings' => new CustomFieldSettingsData(
-            visibility: new VisibilityData(
-                mode: VisibilityMode::SHOW_WHEN,
-                conditions: new DataCollection(VisibilityConditionData::class, [
-                    new VisibilityConditionData(
-                        field_code: $status->code,
-                        operator: VisibilityOperator::EQUALS,
-                        value: 'Done',
-                        source: ConditionSource::CustomField,
-                    ),
-                ]),
-            ),
-        ),
-    ]);
-}
-
-it('renders nothing when no active proposal id is set', function (): void {
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->assertSet('pendingActionId', null)
-        ->assertDontSee('Acme Corp');
-});
-
-it('loads and renders the active pending action summary', function (): void {
-    $action = proposalCardPa($this->user,
-        ['name' => 'Acme Corp'],
-        ['title' => 'Create Company', 'summary' => 'Create company "Acme Corp"', 'fields' => [['label' => 'Name', 'value' => 'Acme Corp']]],
-    );
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->assertSet('pendingActionId', $action->getKey())
-        ->assertSeeHtml('data-proposal-record-chip')
-        ->assertSeeHtml('data-record-type="company"')
-        ->assertSee('Acme Corp');
-});
-
-it('refuses a pending action from another tenant', function (): void {
-    $other = User::factory()->withPersonalTeam()->create();
-    $foreign = proposalCardPa($other, ['name' => 'Foreign'], ['title' => 'x', 'summary' => 'x', 'fields' => []]);
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $foreign->getKey(), context: 'conversation')
-        ->assertSet('pendingActionId', null);
-});
-
-it('ignores set-active events targeted at a different chat context', function (): void {
-    $action = proposalCardPa($this->user, ['name' => 'Acme'], ['title' => 't', 'summary' => 's', 'fields' => []]);
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'side-panel')
-        ->assertSet('pendingActionId', null);
-});
-
-it('focuses a clicked batch record and snaps a resolved index to the first unresolved', function (): void {
-    $action = makeBatchCompanyProposal($this->user, ['Alpha', 'Beta', 'Gamma']);
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->assertSet('cursor', 0)
-        ->call('focusItem', (string) $action->getKey(), 2)->assertSet('cursor', 2)
-        ->call('focusItem', (string) $action->getKey(), 1)->assertSet('cursor', 1);
-
-    $action->update(['result_data' => ['items' => ['0' => ['status' => 'approved']]]]);
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->call('focusItem', (string) $action->getKey(), 0)
-        ->assertSet('cursor', 1);
-});
-
-it('starts the cursor at the first unresolved record', function (): void {
-    $action = makeBatchCompanyProposal($this->user, ['Alpha', 'Beta', 'Gamma']);
-    $action->update(['result_data' => ['items' => ['0' => ['status' => 'approved']]]]);
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->assertSet('cursor', 1);
-});
-
-it('does not surface an expired pending action', function (): void {
-    $action = proposalCardPa($this->user, ['name' => 'Stale'], ['title' => 't', 'summary' => 's', 'fields' => []]);
-    $action->update(['expires_at' => now()->subMinute()]);
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->assertSet('pendingActionId', null);
-});
-
-it('creates only the active batch record and advances to the next', function (): void {
-    Bus::fake();
-    $action = makeBatchCompanyProposal($this->user, ['Alpha', 'Beta']);
-
-    $component = Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->call('createCurrent')
-        ->assertDispatched('proposal:resolved');
-
-    // Only the record on screen was committed; the card advanced to Beta.
-    expect(Company::query()->where('team_id', $this->team->getKey())->pluck('name')->all())->toBe(['Alpha']);
-    expect($action->fresh()->status)->toBe(PendingActionStatus::Pending);
-    $component->assertSet('cursor', 1);
-
-    $component->call('createCurrent')
-        ->assertSet('pendingActionId', null);
-
-    expect(Company::query()->where('team_id', $this->team->getKey())->orderBy('name')->pluck('name')->all())
-        ->toBe(['Alpha', 'Beta']);
-    expect($action->fresh()->status)->toBe(PendingActionStatus::Approved);
-});
-
-it('creates the single proposal record and collapses the dock', function (): void {
-    Bus::fake();
-    $action = proposalCardPa($this->user,
-        ['name' => 'Acme Corp'],
-        ['title' => 'Create Company', 'summary' => 'Create company "Acme Corp"', 'fields' => [['label' => 'Name', 'value' => 'Acme Corp']]],
-    );
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->call('createCurrent')
-        ->assertDispatched('proposal:resolved')
-        ->assertSet('pendingActionId', null);
-
-    expect(Company::query()->where('team_id', $this->team->getKey())->where('name', 'Acme Corp')->exists())->toBeTrue();
-    expect($action->fresh()->status)->toBe(PendingActionStatus::Approved);
-});
-
-it('finalizes the batch on the last item and collapses the dock', function (): void {
-    Bus::fake();
-    $action = makeBatchCompanyProposal($this->user, ['Alpha', 'Beta']);
-    $action->update(['result_data' => ['items' => ['0' => ['status' => 'approved', 'id' => 'x']], 'ids' => ['x']]]);
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->assertSet('cursor', 1)
-        ->call('createCurrent')
-        ->assertDispatched('proposal:resolved')
-        ->assertSet('pendingActionId', null);
-
-    expect($action->fresh()->status)->toBe(PendingActionStatus::Approved);
-});
-
-it('discards only the active batch record and advances to the next', function (): void {
-    Bus::fake();
-    $action = makeBatchCompanyProposal($this->user, ['Alpha', 'Beta']);
-
-    $component = Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->call('discardCurrent')
-        ->assertDispatched('proposal:resolved');
-
-    expect($action->fresh()->status)->toBe(PendingActionStatus::Pending);
-    $component->assertSet('cursor', 1);
-
-    $component->call('discardCurrent')
-        ->assertSet('pendingActionId', null);
-
-    expect(Company::query()->where('team_id', $this->team->getKey())->count())->toBe(0);
-    expect($action->fresh()->status)->toBe(PendingActionStatus::Rejected);
-});
-
-it('finalizes after a skip plus create-all without dispatching a continuation', function (): void {
-    Bus::fake();
-    $action = makeBatchCompanyProposal($this->user, ['Alpha', 'Beta']);
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->call('skipItem', (string) $action->getKey(), 1)
-        ->call('createCurrent')
-        ->assertSet('pendingActionId', null);
-    expect($action->fresh()->status)->not->toBe(PendingActionStatus::Pending);
-
-    expect(Company::query()->where('team_id', $this->team->getKey())->pluck('name')->all())
-        ->toBe(['Alpha']);
-});
-
-it('marks a fully-discarded batch as rejected', function (): void {
-    Bus::fake();
-    $action = makeBatchCompanyProposal($this->user, ['Alpha', 'Beta']);
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->call('discardCurrent')
-        ->call('discardCurrent');
-
-    expect($action->fresh()->status)->toBe(PendingActionStatus::Rejected);
-    expect(Company::query()->where('team_id', $this->team->getKey())->count())->toBe(0);
-});
-
-it('emits proposal:resolve-failed and does not advance when the service rejects the resolution', function (): void {
-    Bus::fake();
-    $action = makeBatchCompanyProposal($this->user, ['Alpha', 'Beta']);
-
-    // A non-allowlisted action_class makes the service reject the approve for a valid
-    // (in-range, unresolved) item — exercising the catch -> resolve-failed path rather
-    // than the stale-cursor guard.
-    $action->update(['action_class' => 'stdClass']);
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->call('createCurrent') // cursor 0 is a valid unresolved item; the service throws
-        ->assertDispatched('proposal:resolve-failed')
-        ->assertNotDispatched('proposal:resolved')
-        ->assertSet('pendingActionId', $action->getKey()); // not cleared
-
-    expect($action->fresh()->status)->toBe(PendingActionStatus::Pending);
-    expect(Company::query()->where('team_id', $this->team->getKey())->count())->toBe(0);
-});
-
-it('drops a skipped item from the dock queue and cannot re-decide it', function (): void {
-    Bus::fake();
-    $action = makeBatchCompanyProposal($this->user, ['Alpha', 'Beta', 'Gamma']);
-
-    $component = Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->call('skipItem', (string) $action->getKey(), 0);
-
-    // Alpha left the queue: two records remain and the cursor sits on Beta.
-    $component->assertViewHas('remainingCount', 2)
-        ->assertSet('cursor', 1);
-
-    // Skipping the already-decided index again is a no-op, not a re-run.
-    $component->call('skipItem', (string) $action->getKey(), 0)
-        ->assertSet('cursor', 1);
-
-    // Each per-record Create commits only the record on screen.
-    $component->call('createCurrent')
-        ->call('createCurrent');
-
-    expect(Company::query()->where('team_id', $this->team->getKey())->orderBy('name')->pluck('name')->all())
-        ->toBe(['Beta', 'Gamma']);
-    expect($action->fresh()->status)->toBe(PendingActionStatus::Approved);
-});
-
-it('does nothing when createCurrent is called while a field edit is open', function (): void {
-    Bus::fake();
-    $action = makeBatchCompanyProposal($this->user, ['Alpha', 'Beta']);
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->call('editField', 'name')
-        ->call('createCurrent')
-        ->assertNotDispatched('proposal:resolved');
-
-    expect(Company::query()->where('team_id', $this->team->getKey())->count())->toBe(0);
-});
-
-it('routes the create-current shortcut to the active batch record for the matching context', function (): void {
-    Bus::fake();
-    $action = makeBatchCompanyProposal($this->user, ['Alpha', 'Beta']);
-
-    $component = Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->dispatch('proposal:create-current', context: 'conversation');
-
-    expect(Company::query()->where('team_id', $this->team->getKey())->pluck('name')->all())->toBe(['Alpha']);
-
-    $component->dispatch('proposal:create-current', context: 'conversation')
-        ->assertSet('pendingActionId', null);
-
-    expect(Company::query()->where('team_id', $this->team->getKey())->orderBy('name')->pluck('name')->all())
-        ->toBe(['Alpha', 'Beta']);
-});
-
-it('ignores the create-current shortcut for a different context', function (): void {
-    Bus::fake();
-    $action = makeBatchCompanyProposal($this->user, ['Alpha', 'Beta']);
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->dispatch('proposal:create-current', context: 'side-panel')
-        ->assertSet('cursor', 0);
-
-    expect(Company::query()->where('team_id', $this->team->getKey())->count())->toBe(0);
-});
-
-it('paginates a batch one record at a time with a per-record footer', function (): void {
-    $action = makeBatchCompanyProposal($this->user, ['Alpha', 'Beta', 'Gamma']);
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->assertSee('Alpha')
-        ->assertDontSee('Beta')
-        ->assertDontSee('Gamma')
-        ->assertSee('1/3')
-        ->assertSee('Create')
-        ->assertSee('Discard')
-        ->assertDontSee('Create all')
-        ->call('nextItem')
-        ->assertSee('Beta')
-        ->assertDontSee('Alpha')
-        ->assertSee('2/3');
-});
-
-it('clamps pagination at both ends of the undecided records', function (): void {
-    $action = makeBatchCompanyProposal($this->user, ['Alpha', 'Beta']);
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->call('prevItem')
-        ->assertSet('cursor', 0)
-        ->call('nextItem')
-        ->assertSet('cursor', 1)
-        ->call('nextItem')
-        ->assertSet('cursor', 1)
-        ->call('prevItem')
-        ->assertSet('cursor', 0);
-});
-
-it('excludes an unchecked field from the write and records it in result_data', function (): void {
-    Bus::fake();
-    $action = proposalCardPa($this->user,
-        ['name' => 'Acme Corp', 'account_owner_id' => $this->user->getKey()],
-        ['title' => 'Create Company', 'summary' => 'Create company "Acme Corp"', 'fields' => [
-            ['label' => 'Name', 'code' => 'name', 'value' => 'Acme Corp'],
-            ['label' => 'Account owner', 'code' => 'account_owner_id', 'value' => 'Someone'],
-        ]],
-    );
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->call('toggleField', 'account_owner_id')
-        ->assertSet('excludedFields', ['account_owner_id'])
-        ->call('createCurrent')
-        ->assertDispatched('proposal:resolved');
-
-    $company = Company::query()->where('team_id', $this->team->getKey())->where('name', 'Acme Corp')->first();
-
-    expect($company)->not->toBeNull()
-        ->and($company->account_owner_id)->toBeNull()
-        ->and($action->fresh()->result_data['excluded'] ?? null)->toBe(['account_owner_id'])
-        ->and($action->fresh()->action_data)->toHaveKey('account_owner_id');
-});
-
-it('never lets the title field be unchecked', function (): void {
-    $action = proposalCardPa($this->user,
-        ['name' => 'Acme Corp', 'account_owner_id' => $this->user->getKey()],
-        ['title' => 'Create Company', 'summary' => 'Create company "Acme Corp"', 'fields' => [
-            ['label' => 'Name', 'code' => 'name', 'value' => 'Acme Corp'],
-        ]],
-    );
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->call('toggleField', 'name')
-        ->assertSet('excludedFields', []);
-});
-
-it('applies exclusions per batch record and resets them on every navigation', function (): void {
-    Bus::fake();
-    $records = [
-        ['name' => 'Alpha', 'account_owner_id' => $this->user->getKey()],
-        ['name' => 'Beta', 'account_owner_id' => $this->user->getKey()],
-    ];
-    $items = array_map(static fn (array $record): array => [
-        'title' => $record['name'],
-        'summary' => "Create company \"{$record['name']}\"",
-        'fields' => [
-            ['label' => 'Name', 'code' => 'name', 'value' => $record['name']],
-            ['label' => 'Account owner', 'code' => 'account_owner_id', 'value' => 'Someone'],
-        ],
-    ], $records);
-    $action = proposalCardPa(
-        $this->user,
-        ['_batch' => true, 'records' => $records],
-        ['title' => 'Create Companies', 'summary' => 'Create 2 companies', 'items' => $items],
-    );
-
-    $component = Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->call('toggleField', 'account_owner_id')
-        ->assertSet('excludedFields', ['account_owner_id']);
-
-    // Paging away and back clears the exclusion: it belonged to Alpha's card.
-    $component->call('nextItem')
-        ->assertSet('excludedFields', [])
-        ->call('prevItem')
-        ->call('toggleField', 'account_owner_id')
-        ->call('createCurrent')
-        ->call('createCurrent');
-
-    $owners = Company::query()
-        ->where('team_id', $this->team->getKey())
-        ->orderBy('name')
-        ->pluck('account_owner_id', 'name')
-        ->all();
-
-    expect($owners['Alpha'])->toBeNull()
-        ->and($owners['Beta'])->not->toBeNull()
-        ->and($action->fresh()->result_data['items'][0]['excluded'] ?? null)->toBe(['account_owner_id'])
-        ->and($action->fresh()->result_data['items'][1]['excluded'] ?? null)->toBeNull();
-});
-
-it('unchecks and rechecks everything through the master toggle', function (): void {
-    $action = proposalCardPa($this->user,
-        ['name' => 'Acme Corp', 'account_owner_id' => $this->user->getKey()],
-        ['title' => 'Create Company', 'summary' => 'Create company "Acme Corp"', 'fields' => [
-            ['label' => 'Name', 'code' => 'name', 'value' => 'Acme Corp'],
-            ['label' => 'Account owner', 'code' => 'account_owner_id', 'value' => 'Someone'],
-        ]],
-    );
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->call('toggleAllFields')
-        ->assertSet('excludedFields', ['account_owner_id'])
-        ->call('toggleAllFields')
-        ->assertSet('excludedFields', []);
-});
-
-it('refuses to approve an update whose every change is unchecked', function (): void {
-    Bus::fake();
-    $company = Company::factory()->for($this->team)->create(['name' => 'Stable Inc', 'account_owner_id' => null]);
-
-    $action = PendingAction::query()->create([
-        'team_id' => $this->team->getKey(),
-        'user_id' => $this->user->getKey(),
-        'conversation_id' => null,
-        'action_class' => 'App\\Actions\\Company\\UpdateCompany',
-        'operation' => PendingActionOperation::Update,
-        'entity_type' => 'company',
-        'action_data' => ['_record_id' => (string) $company->getKey(), '_model_class' => Company::class, 'account_owner_id' => $this->user->getKey()],
-        'display_data' => [
-            'title' => 'Update Company',
-            'summary' => 'Update company "Stable Inc"',
-            'fields' => [['label' => 'Account owner', 'code' => 'account_owner_id', 'old' => 'Nobody', 'new' => 'Someone']],
-        ],
-        'status' => PendingActionStatus::Pending,
-        'expires_at' => now()->addMinutes(15),
-    ]);
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->call('toggleField', 'account_owner_id')
-        ->call('createCurrent')
-        ->assertNotDispatched('proposal:resolved');
-
-    expect($action->fresh()->status)->toBe(PendingActionStatus::Pending)
-        ->and($company->fresh()->account_owner_id)->toBeNull();
-});
-
-it('renders a single (non-batch) proposal without a stepper', function (): void {
-    $action = proposalCardPa($this->user, ['name' => 'Solo Inc'], ['title' => 'Create Company', 'summary' => 'Solo Inc', 'fields' => [['label' => 'Name', 'value' => 'Solo Inc']]]);
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->assertSee('Solo Inc')
-        ->assertSee('Create');
-});
-
-it('builds the real custom-field component for the edited field, prefilled from action_data', function (): void {
-    [$field, $optionIds] = seedTaskSingleChoiceField($this->team);
-    $action = makeTaskProposal($this->user, ['title' => 'Edit me', 'custom_fields' => [$field->code => $optionIds[0]]]);
-
-    $component = Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->call('editField', $field->code)
-        ->assertSet('editingFieldCode', $field->code)
-        ->assertSet("data.custom_fields.{$field->code}", $optionIds[0])
-        ->assertHasNoErrors();
-
-    $expectedName = "custom_fields.{$field->code}";
-    $flat = $component->instance()->form->getFlatComponents();
-    $built = collect($flat)->first(fn (mixed $c): bool => $c instanceof Field && $c->getName() === $expectedName);
-
-    expect($built)->not->toBeNull('the scoped Filament custom-field component should be built into the form');
-});
-
-it('does not throw building a field with a cross-field visibility condition (fails open under ->only())', function (): void {
-    $dependent = seedTaskFieldWithVisibilityCondition($this->team);
-    $action = makeTaskProposal($this->user, ['title' => 'T', 'custom_fields' => []]);
-
-    $component = Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->call('editField', $dependent->code)
-        ->assertSet('editingFieldCode', $dependent->code)
-        ->assertHasNoErrors();
-
-    $flat = $component->instance()->form->getFlatComponents();
-    $built = collect($flat)->first(fn (mixed $c): bool => $c instanceof Field
-        && $c->getName() === "custom_fields.{$dependent->code}");
-
-    expect($built)->not->toBeNull('the field with a sibling visibility condition should still build under ->only()');
-});
-
-it('saves an edited custom field through ProposalEditor without executing', function (): void {
-    Bus::fake();
-    [$field, $optionIds] = seedTaskSingleChoiceField($this->team);
-    $action = makeTaskProposal($this->user, ['title' => 'T', 'custom_fields' => [$field->code => $optionIds[0]]]);
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->call('editField', $field->code)
-        ->set("data.custom_fields.{$field->code}", $optionIds[1])
-        ->call('saveField')
-        ->assertSet('editingFieldCode', null)
-        ->assertHasNoErrors();
-
-    $fresh = $action->fresh();
-    expect($fresh->status)->toBe(PendingActionStatus::Pending)
-        ->and($fresh->action_data['custom_fields'][$field->code])->toBe($optionIds[1]);
-    expect(Task::query()->where('team_id', $this->team->getKey())->count())->toBe(0);
-});
-
-it('preserves other custom fields when only one is edited', function (): void {
-    Bus::fake();
-    [$status, $statusOptionIds] = seedTaskSingleChoiceField($this->team);
-
-    $priority = CustomField::query()
-        ->where('tenant_id', $this->team->getKey())
-        ->where('entity_type', 'task')
-        ->where('code', 'priority')
-        ->with('options')
-        ->first();
-    expect($priority)->not->toBeNull('seeded task priority field is required for this test');
-    $priorityOptionId = (string) $priority->options->first()->id;
-
-    $action = makeTaskProposal($this->user, [
-        'title' => 'T',
-        'custom_fields' => [
-            $status->code => $statusOptionIds[0],
-            $priority->code => $priorityOptionId,
-        ],
-    ]);
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->call('editField', $status->code)
-        ->set("data.custom_fields.{$status->code}", $statusOptionIds[1])
-        ->call('saveField')
-        ->assertSet('editingFieldCode', null)
-        ->assertHasNoErrors();
-
-    $fresh = $action->fresh();
-    expect($fresh->action_data['custom_fields'][$status->code])->toBe($statusOptionIds[1])
-        ->and($fresh->action_data['custom_fields'][$priority->code])->toBe($priorityOptionId);
-});
-
-it('cancels an inline edit without persisting and leaves action_data untouched', function (): void {
-    [$field, $optionIds] = seedTaskSingleChoiceField($this->team);
-    $action = makeTaskProposal($this->user, ['title' => 'Keep me', 'custom_fields' => [$field->code => $optionIds[0]]]);
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->call('editField', $field->code)
-        ->set("data.custom_fields.{$field->code}", $optionIds[1]) // change the working value...
-        ->call('cancelField')                                     // ...then cancel
-        ->assertSet('editingFieldCode', null);
-
-    expect($action->fresh()->action_data['custom_fields'][$field->code])->toBe($optionIds[0]);
-});
-
-it('edits a core text field (title) in place and persists it via applyEdit without executing', function (): void {
-    Bus::fake();
-    $action = makeTaskProposal($this->user, ['title' => 'Old Title']);
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->call('editField', 'title')
-        ->assertSet('editingFieldCode', 'title')
-        ->assertSet('data.title', 'Old Title')
-        ->set('data.title', 'New Title')
-        ->call('saveField')
-        ->assertSet('editingFieldCode', null)
-        ->assertHasNoErrors();
-
-    $fresh = $action->fresh();
-    expect($fresh->status)->toBe(PendingActionStatus::Pending)
-        ->and($fresh->action_data['title'])->toBe('New Title');
-    expect(Task::query()->where('team_id', $this->team->getKey())->count())->toBe(0);
-});
-
-it('rejects an out-of-options choice value at the form layer without persisting', function (): void {
-    Bus::fake();
-    [$field, $optionIds] = seedTaskSingleChoiceField($this->team);
-    $action = makeTaskProposal($this->user, ['title' => 'T', 'custom_fields' => [$field->code => $optionIds[0]]]);
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->call('editField', $field->code)
-        ->set("data.custom_fields.{$field->code}", 'not-an-option-id')
-        ->call('saveField')
-        ->assertSet('editingFieldCode', $field->code)
-        ->assertHasErrors("data.custom_fields.{$field->code}");
-
-    expect($action->fresh()->action_data['custom_fields'][$field->code])->toBe($optionIds[0]);
-});
-
-it('rejects an empty required core name at the form layer and keeps the proposal pending', function (): void {
-    Bus::fake();
-    $action = proposalCardPa(
-        $this->user,
-        ['name' => 'Acme Corp', 'account_owner_id' => (string) $this->user->getKey()],
-        ['title' => 'Create Company', 'summary' => 'Acme Corp', 'fields' => [['label' => 'Name', 'value' => 'Acme Corp']]],
-    );
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->call('editField', 'name')
-        ->set('data.name', '   ')
-        ->call('saveField')
-        ->assertSet('editingFieldCode', 'name')
-        ->assertHasErrors('data.name');
-
-    $fresh = $action->fresh();
-    expect($fresh->status)->toBe(PendingActionStatus::Pending)
-        ->and($fresh->action_data['name'])->toBe('Acme Corp');
-});
-
-it('exposes editable codes for the entity (core keys + non-deferred custom fields)', function (): void {
-    [$field] = seedTaskSingleChoiceField($this->team);
-    $action = makeTaskProposal($this->user, ['title' => 'T', 'custom_fields' => [$field->code => null]]);
-
-    $codes = Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->instance()->editableCodes();
-
-    expect($codes)->toContain('title')
-        ->and($codes)->toContain($field->code);
-});
-
-it('omits deferred custom fields (file upload, record lookup) from the editable codes', function (): void {
-    $fileField = CustomField::query()->create([
-        'tenant_id' => $this->team->getKey(),
-        'entity_type' => 'task',
-        'code' => 'attachment',
-        'name' => 'Attachment',
-        'type' => CustomFieldType::FILE_UPLOAD->value,
-        'sort_order' => 90,
-        'validation_rules' => [],
-        'active' => true,
-        'system_defined' => false,
-    ]);
-
-    // A record-lookup field resolves to a MULTI_CHOICE data type, so kindFor()
-    // would otherwise admit it as a 'multiselect' — only isDeferred()'s RECORD /
-    // lookup_type branch keeps it out. This is the row that makes the deferral
-    // load-bearing (the file-upload type is disabled in config, so it is excluded
-    // by the kindFor() fallback regardless).
-    $recordField = CustomField::query()->create([
-        'tenant_id' => $this->team->getKey(),
-        'entity_type' => 'task',
-        'code' => 'related_company',
-        'name' => 'Related Company',
-        'type' => CustomFieldType::RECORD->value,
-        'lookup_type' => 'company',
-        'sort_order' => 91,
-        'validation_rules' => [],
-        'active' => true,
-        'system_defined' => false,
-    ]);
-
-    $action = makeTaskProposal($this->user, ['title' => 'T', 'custom_fields' => []]);
-
-    $codes = Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->instance()->editableCodes();
-
-    expect($codes)->toContain('title')
-        ->and($codes)->not->toContain($fileField->code)
-        ->and($codes)->not->toContain($recordField->code);
-});
-
-it('rebuilds the current record fields with codes on editable rows and no divergence from stored display', function (): void {
-    [$field, $optionIds] = seedTaskSingleChoiceField($this->team);
-    $optionLabel = (string) $field->options->firstWhere('id', $optionIds[0])->name;
-
-    $tool = resolve(CreateTaskTool::class);
-    $tool->setConversationId(null);
-    $tool->handle(new Request(['records' => [['title' => 'My Task', 'custom_fields' => [$field->code => $optionLabel]]]]));
-
-    $action = PendingAction::query()->latest()->firstOrFail();
-
-    $fields = Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->instance()->currentRecordFields();
-
-    $titleRow = collect($fields)->firstWhere('label', 'Title');
-    expect($titleRow['code'] ?? null)->toBe('title');
-
-    $customRow = collect($fields)->firstWhere('code', $field->code);
-    expect($customRow)->not->toBeNull();
-
-    $stored = $action->display_data['fields'] ?? ($action->display_data['items'][0]['fields'] ?? []);
-
-    expect(collect($fields)->pluck('label')->all())->toBe(collect($stored)->pluck('label')->all())
-        ->and(collect($fields)->pluck('value')->all())->toBe(collect($stored)->pluck('value')->all())
-        ->and(collect($fields)->pluck('new')->all())->toBe(collect($stored)->pluck('new')->all());
-});
-
-it('edits a custom field on a batch item without touching sibling records', function (): void {
-    Bus::fake();
-    [$field, $optionIds] = seedTaskSingleChoiceField($this->team);
-    $action = makeBatchTaskProposal($this->user, [
-        ['title' => 'Task A', 'custom_fields' => [$field->code => $optionIds[0]]],
-        ['title' => 'Task B', 'custom_fields' => [$field->code => $optionIds[0]]],
-    ]);
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->call('focusItem', (string) $action->getKey(), 1)
-        ->assertSet('cursor', 1)
-        ->call('editField', $field->code)
-        ->assertSet("data.custom_fields.{$field->code}", $optionIds[0])
-        ->set("data.custom_fields.{$field->code}", $optionIds[1])
-        ->call('saveField')
-        ->assertSet('editingFieldCode', null)
-        ->assertHasNoErrors();
-
-    $records = array_values($action->fresh()->action_data['records']);
-    expect($records[1]['custom_fields'][$field->code])->toBe($optionIds[1])
-        ->and($records[0]['custom_fields'][$field->code])->toBe($optionIds[0]);
-    expect($action->fresh()->status)->toBe(PendingActionStatus::Pending);
-    expect(Task::query()->where('team_id', $this->team->getKey())->count())->toBe(0);
-});
-
-it('shows an edit affordance for an editable field and renders the inline editor when editing', function (): void {
-    [$field, $optionIds] = seedTaskSingleChoiceField($this->team);
-    $action = makeTaskProposal($this->user, ['title' => 'T', 'custom_fields' => [$field->code => $optionIds[0]]]);
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->assertSeeHtml('editField')
-        ->call('editField', 'title')
-        ->assertSeeHtml('wire:click="saveField"')
-        ->assertSeeHtml('wire:click="cancelField"');
-});
-
-it('rebuilds a company record (with account owner + custom field) without diverging from stored display', function (): void {
-    $linkedin = CustomField::query()
-        ->where('tenant_id', $this->team->getKey())
-        ->where('entity_type', 'company')
-        ->where('code', 'linkedin')
-        ->first();
-
-    expect($linkedin)->not->toBeNull('seeded company linkedin field is required for this test');
-
-    $tool = resolve(CreateCompanyTool::class);
-    $tool->setConversationId(null);
-    $tool->handle(new Request(['records' => [[
-        'name' => 'Acme Corp',
-        'account_owner_id' => (string) $this->user->getKey(),
-        'custom_fields' => ['linkedin' => ['https://linkedin.com/company/acme']],
-    ]]]));
-
-    $action = PendingAction::query()->latest()->firstOrFail();
-
-    $fields = Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->instance()->currentRecordFields();
-
-    $nameRow = collect($fields)->firstWhere('label', 'Name');
-    expect($nameRow['code'] ?? null)->toBe('name');
-
-    $ownerRow = collect($fields)->firstWhere('label', 'Account Owner');
-    expect($ownerRow)->not->toBeNull()
-        ->and($ownerRow['code'] ?? null)->toBe('account_owner_id')
-        ->and($ownerRow['value'] ?? null)->toBe($this->user->name);
-
-    $customRow = collect($fields)->firstWhere('code', 'linkedin');
-    expect($customRow)->not->toBeNull();
-
-    $stored = $action->display_data['fields'] ?? ($action->display_data['items'][0]['fields'] ?? []);
-
-    expect(collect($fields)->pluck('label')->all())->toBe(collect($stored)->pluck('label')->all())
-        ->and(collect($fields)->pluck('value')->all())->toBe(collect($stored)->pluck('value')->all())
-        ->and(collect($fields)->pluck('new')->all())->toBe(collect($stored)->pluck('new')->all());
-});
-
-it('ignores skipItem while a field edit is open', function (): void {
-    Bus::fake();
-    $action = makeBatchCompanyProposal($this->user, ['Alpha', 'Beta']);
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->call('editField', 'name')
-        ->call('skipItem', (string) $action->getKey(), 1)
-        ->assertNotDispatched('proposal:resolved');
-
-    expect($action->fresh()->result_data)->toBeNull();
-});
-
-it('shrinks the pagination count as records are skipped and hides it for the last one', function (): void {
-    Bus::fake();
-    $action = makeBatchCompanyProposal($this->user, ['Alpha', 'Beta', 'Gamma']);
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->assertSee('1/3')
-        ->call('skipItem', (string) $action->getKey(), 2)
-        ->assertSee('1/2')
-        ->assertDontSee('Gamma')
-        ->call('skipItem', (string) $action->getKey(), 1)
-        ->assertSee('Create')
-        ->assertDontSee('1/1');
-});
-
-it('keeps a batch open when the active record fails and keeps the earlier commits', function (): void {
-    Bus::fake();
-    $company = Company::factory()->for($this->team)->create(['name' => 'Old name']);
-
-    $action = PendingAction::query()->create([
-        'team_id' => $this->team->getKey(),
-        'user_id' => $this->user->getKey(),
-        'conversation_id' => null,
-        'action_class' => 'App\\Actions\\Company\\UpdateCompany',
-        'operation' => PendingActionOperation::Update,
-        'entity_type' => 'company',
-        'action_data' => [
-            '_batch' => true,
-            'records' => [
-                ['_record_id' => (string) $company->getKey(), '_model_class' => Company::class, 'name' => 'New name'],
-                ['_record_id' => '01J00000000000000000000000', '_model_class' => Company::class, 'name' => 'Ghost'],
-            ],
-        ],
-        'display_data' => [
-            'title' => 'Update 2 companies',
-            'summary' => 'Update 2 companies',
-            'items' => [
-                ['summary' => 'Update company "Old name"', 'fields' => []],
-                ['summary' => 'Update company "Ghost"', 'fields' => []],
-            ],
-        ],
-        'status' => PendingActionStatus::Pending,
-        'expires_at' => now()->addMinutes(15),
-    ]);
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->call('createCurrent')
-        ->assertDispatched('proposal:resolved')
-        ->call('createCurrent')
-        ->assertDispatched('proposal:resolve-failed');
-
-    // The first record committed; the vanished one failed its own approval
-    // without undoing it, and the batch stays open on the failing record.
-    expect($company->fresh()->name)->toBe('New name')
-        ->and($action->fresh()->status)->toBe(PendingActionStatus::Pending)
-        ->and($action->fresh()->result_data['items'][0]['status'] ?? null)->toBe('approved');
-});
-
-it('resolves a delete batch per item through the component, deleting only approved records', function (): void {
-    Bus::fake();
-    $tasks = Task::factory()->count(2)->for($this->team)->create();
-
-    $action = PendingAction::query()->create([
-        'team_id' => $this->team->getKey(),
-        'user_id' => $this->user->getKey(),
-        'conversation_id' => null,
-        'action_class' => 'App\\Actions\\Task\\DeleteTask',
-        'operation' => PendingActionOperation::Delete,
-        'entity_type' => 'task',
-        'action_data' => [
-            '_batch' => true,
-            'records' => $tasks->map(fn (Task $t): array => ['_record_id' => $t->getKey(), '_model_class' => Task::class])->all(),
-        ],
-        'display_data' => [
-            'title' => 'Delete 2 Tasks',
-            'summary' => 'Delete 2 tasks',
-            'items' => $tasks->map(fn (Task $t): array => [
-                'summary' => "Delete Task \"{$t->title}\"",
-                'fields' => [['label' => 'Name', 'value' => $t->title]],
-            ])->all(),
-        ],
-        'status' => PendingActionStatus::Pending,
-        'expires_at' => now()->addMinutes(15),
-    ]);
-
-    $component = Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->call('skipItem', (string) $action->getKey(), 1); // keep task 1
-
-    // The batch is not yet finalized: item 1 is skipped, item 0 still pending.
-    expect(Task::query()->whereKey($tasks[1]->getKey())->exists())->toBeTrue()
-        ->and(Task::query()->whereKey($tasks[0]->getKey())->exists())->toBeTrue()
-        ->and($action->fresh()->status)->toBe(PendingActionStatus::Pending);
-
-    $component->call('createCurrent'); // delete the remaining item 0 -> finalize
-
-    expect($action->fresh()->status)->toBe(PendingActionStatus::Approved)
-        ->and(Task::query()->whereKey($tasks[0]->getKey())->exists())->toBeFalse()
-        ->and(Task::query()->whereKey($tasks[1]->getKey())->exists())->toBeTrue();
-});
-
-it('offers no inline-edit codes for a delete proposal', function (): void {
-    $task = Task::factory()->for($this->team)->create();
-
-    $action = PendingAction::query()->create([
-        'team_id' => $this->team->getKey(),
-        'user_id' => $this->user->getKey(),
-        'conversation_id' => null,
-        'action_class' => 'App\\Actions\\Task\\DeleteTask',
-        'operation' => PendingActionOperation::Delete,
-        'entity_type' => 'task',
-        'action_data' => ['_record_ids' => [$task->getKey()], '_model_class' => Task::class],
-        'display_data' => ['title' => 'Delete Task', 'summary' => "Delete Task \"{$task->title}\"", 'fields' => [['label' => 'Name', 'value' => $task->title]]],
-        'status' => PendingActionStatus::Pending,
-        'expires_at' => now()->addMinutes(15),
-    ]);
-
-    $codes = Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->instance()->editableCodes();
-
-    expect($codes)->toBe([]);
-});
-
-it('surfaces a failure when the assignee left the workspace between proposal and approval', function (): void {
-    $member = User::factory()->create();
-    $this->team->users()->attach($member, ['role' => 'editor']);
-
-    $action = PendingAction::query()->create([
-        'team_id' => $this->team->getKey(),
-        'user_id' => $this->user->getKey(),
-        'conversation_id' => null,
-        'action_class' => 'App\\Actions\\Task\\CreateTask',
-        'operation' => PendingActionOperation::Create,
-        'entity_type' => 'task',
-        'action_data' => ['title' => 'Follow up call', 'assignee_ids' => [(string) $member->getKey()]],
-        'display_data' => ['title' => 'Create Task', 'summary' => 'Create task "Follow up call"', 'fields' => []],
-        'status' => PendingActionStatus::Pending,
-        'expires_at' => now()->addMinutes(15),
-    ]);
-
-    $this->team->users()->detach($member);
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->call('createCurrent')
-        ->assertDispatched('proposal:resolve-failed')
-        ->assertHasErrors('resolve');
-
-    expect(Task::query()->where('title', 'Follow up call')->exists())->toBeFalse()
-        ->and($action->fresh()->status)->toBe(PendingActionStatus::Pending)
-        // The reason survives on the row, so the model can explain the failure
-        // after the proposal is eventually decided.
-        ->and($action->fresh()->result_data['last_error'] ?? null)->toBeString()->not->toBe('');
-});
-
-it('renders the resolve failure in the dock so the approval is never a silent no-op', function (): void {
-    $action = PendingAction::query()->create([
-        'team_id' => $this->team->getKey(),
-        'user_id' => $this->user->getKey(),
-        'conversation_id' => null,
-        'action_class' => 'App\\Actions\\Task\\CreateTask',
-        'operation' => PendingActionOperation::Create,
-        'entity_type' => 'task',
-        'action_data' => ['title' => 'Ghost assignee', 'assignee_ids' => [(string) User::factory()->create()->getKey()]],
-        'display_data' => ['title' => 'Create Task', 'summary' => 'Create task "Ghost assignee"', 'fields' => []],
-        'status' => PendingActionStatus::Pending,
-        'expires_at' => now()->addMinutes(15),
-    ]);
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->call('createCurrent')
-        ->assertSee('not in your workspace');
-});
-
-/**
- * A pending Create-custom-field proposal shaped exactly like CreateCustomFieldTool emits.
- */
-function customFieldProposal(User $user, string $entityType, string $name, string $type): PendingAction
-{
-    return PendingAction::query()->create([
-        'team_id' => $user->currentTeam->getKey(),
-        'user_id' => $user->getKey(),
-        'conversation_id' => null,
-        'action_class' => 'App\\Actions\\CustomFields\\CreateCustomField',
-        'operation' => PendingActionOperation::Create,
-        'entity_type' => 'custom_field',
-        'action_data' => ['entity_type' => $entityType, 'name' => $name, 'type' => $type],
-        'display_data' => [
-            'title' => 'Create Custom Field',
-            'summary' => "Create \"{$name}\" ({$type}) on {$entityType}",
-            'fields' => [
-                ['label' => 'Entity', 'value' => $entityType],
-                ['label' => 'Name', 'value' => $name],
-                ['label' => 'Type', 'value' => $type],
-            ],
-        ],
-        'status' => PendingActionStatus::Pending,
-        'expires_at' => now()->addMinutes(15),
-    ]);
-}
-
-it('approving a custom field proposal dispatches a record link to the management page', function (): void {
-    Bus::fake();
-    $action = customFieldProposal($this->user, 'people', 'Age', 'number');
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->call('createCurrent')
-        ->assertDispatched('proposal:resolved', function (string $event, array $params): bool {
-            $record = $params['record'] ?? null;
-
-            return is_array($record)
-                && $record['type'] === 'custom_field'
-                && $record['label'] === 'Age'
-                && str_contains((string) $record['url'], 'currentEntityType=people');
-        });
-});
-
-it('surfaces a readable error on the card when the field name was taken after the proposal', function (): void {
-    Bus::fake();
-    $action = customFieldProposal($this->user, 'people', 'Age', 'number');
-
-    CustomField::factory()->create([
-        config('custom-fields.database.column_names.tenant_foreign_key') => $this->team->getKey(),
-        'entity_type' => 'people',
-        'name' => 'Age',
-        'code' => 'age',
-        'type' => 'number',
-    ]);
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->call('createCurrent')
-        ->assertDispatched('proposal:resolve-failed')
-        ->assertHasErrors('resolve');
-
-    expect($action->fresh()->status)->toBe(PendingActionStatus::Pending);
-});
-
-it('never puts a database error message on the card or in the transcript', function (): void {
-    Bus::fake();
-
-    // companies.name is NOT NULL and CreateCompany does not validate it, so an
-    // approval carrying a stale/emptied name reaches Postgres and raises a
-    // QueryException — whose getMessage() carries the statement, the failing row
-    // and the connection host.
-    $action = proposalCardPa($this->user, ['name' => null], ['title' => 'Create Company', 'summary' => 'Create a company']);
-
-    $component = Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation')
-        ->call('createCurrent')
-        ->assertDispatched('proposal:resolve-failed')
-        ->assertNotDispatched('proposal:resolved');
-
-    $errors = $component->errors()->get('resolve');
-    $shown = implode(' ', $errors);
-
-    expect($errors)->not->toBeEmpty()
-        ->and($shown)->not->toContain('SQLSTATE')
-        ->and($shown)->not->toContain('insert into')
-        ->and($shown)->not->toContain('Connection: pgsql')
-        ->and($shown)->not->toContain('companies')
-        ->and($shown)->toBe('This change could not be saved. Please try again.');
-
-    expect($action->fresh()->status)->toBe(PendingActionStatus::Pending);
-    expect(Company::query()->where('team_id', $this->team->getKey())->count())->toBe(0);
-});
-
-describe('plan card', function (): void {
-    /**
-     * A chained turn: company, then a person referencing it, then a task
-     * referencing the person. Same shape the assistant produces live.
-     *
-     * @return array{0: PendingAction, 1: PendingAction, 2: PendingAction}
-     */
-    function planCardSteps(User $user): array
-    {
-        $conversationId = '019dfb00-5555-7000-8000-000000000009';
-        $turnId = '01PLANCARDTURNAAAAAAAAAAAA';
-
-        DB::table('agent_conversations')->insert([
-            'id' => $conversationId,
-            'participant_type' => 'user',
-            'participant_id' => (string) $user->getKey(),
-            'team_id' => $user->currentTeam->getKey(),
-            'title' => '',
-            'created_at' => now(),
-            'updated_at' => now(),
-        ]);
-
-        $tool = function (string $class) use ($conversationId, $turnId) {
-            $instance = resolve($class);
-            $instance->setConversationId($conversationId);
-            $instance->setTurnId($turnId);
-
-            return $instance;
-        };
-
-        $tool(CreateCompanyTool::class)->handle(new Request(['records' => [['name' => 'Northwind Traders']]]));
-        $company = PendingAction::query()->where('entity_type', 'company')->latest('id')->firstOrFail();
-
-        $tool(CreatePersonTool::class)->handle(new Request([
-            'records' => [['name' => 'Priya Raman', 'company_id' => PlanReference::to((string) $company->getKey())]],
-        ]));
-        $person = PendingAction::query()->where('entity_type', 'people')->latest('id')->firstOrFail();
-
-        $tool(CreateTaskTool::class)->handle(new Request([
-            'records' => [['title' => 'Call Priya', 'people_ids' => [PlanReference::to((string) $person->getKey())]]],
-        ]));
-        $task = PendingAction::query()->where('entity_type', 'task')->latest('id')->firstOrFail();
-
-        return [$company, $person, $task];
-    }
-
-    it('presents every step of the turn, in order, with its dependency', function (): void {
-        Bus::fake();
-        [$company] = planCardSteps($this->user);
-
-        $steps = Livewire::test(ProposalCard::class)
-            ->call('setActive', $company->getKey())
-            ->instance()
-            ->stepViews();
-
-        expect($steps)->toHaveCount(3)
-            ->and(array_column($steps, 'position'))->toBe([1, 2, 3])
-            ->and(array_column($steps, 'entity_type'))->toBe(['company', 'people', 'task'])
-            ->and(array_column($steps, 'recordLabel'))->toBe(['Northwind Traders', 'Priya Raman', 'Call Priya'])
-            ->and($steps[0]['blockedBy'])->toBe([])
-            ->and($steps[1]['blockedBy'])->toBe([1])
-            ->and($steps[2]['blockedBy'])->toBe([2])
-            ->and($steps[1]['isActive'])->toBeFalse()
-            ->and($steps[0]['isActive'])->toBeTrue();
-    });
-
-    it('creates every record from one approval', function (): void {
-        Bus::fake();
-        [$company] = planCardSteps($this->user);
-
-        Livewire::test(ProposalCard::class)
-            ->call('setActive', $company->getKey())
-            ->call('approveAll')
-            ->assertSet('pendingActionId', null)
-            ->assertHasNoErrors();
-
-        $created = Company::query()->where('name', 'Northwind Traders')->firstOrFail();
-
-        expect(People::query()->where('name', 'Priya Raman')->value('company_id'))->toBe((string) $created->getKey())
-            ->and(Task::query()->where('title', 'Call Priya')->exists())->toBeTrue();
-    });
-
-    it('refuses a step whose dependency is still pending', function (): void {
-        Bus::fake();
-        [$company, $person] = planCardSteps($this->user);
-
-        Livewire::test(ProposalCard::class)
-            ->call('setActive', $company->getKey())
-            ->call('approveStep', $person->getKey())
-            ->assertHasErrors('resolve');
-
-        expect(People::query()->count())->toBe(0)
-            ->and($person->refresh()->status)->toBe(PendingActionStatus::Pending);
-    });
-
-    it('approves one step and keeps the rest waiting', function (): void {
-        Bus::fake();
-        [$company, $person, $task] = planCardSteps($this->user);
-
-        Livewire::test(ProposalCard::class)
-            ->call('setActive', $company->getKey())
-            ->call('approveStep', $company->getKey())
-            ->assertHasNoErrors();
-
-        expect(Company::query()->where('name', 'Northwind Traders')->exists())->toBeTrue()
-            ->and($person->refresh()->status)->toBe(PendingActionStatus::Pending)
-            ->and($task->refresh()->status)->toBe(PendingActionStatus::Pending);
-    });
-
-    it('cancels the dependent steps when one is rejected', function (): void {
-        Bus::fake();
-        [$company, $person, $task] = planCardSteps($this->user);
-
-        Livewire::test(ProposalCard::class)
-            ->call('setActive', $company->getKey())
-            ->call('rejectStep', $person->getKey())
-            ->assertHasNoErrors();
-
-        expect($person->refresh()->status)->toBe(PendingActionStatus::Rejected)
-            ->and($task->refresh()->status)->toBe(PendingActionStatus::Rejected)
-            ->and($company->refresh()->status)->toBe(PendingActionStatus::Pending);
-    });
-
-    it('discards the whole plan without writing anything', function (): void {
-        Bus::fake();
-        [$company, $person, $task] = planCardSteps($this->user);
-
-        Livewire::test(ProposalCard::class)
-            ->call('setActive', $company->getKey())
-            ->call('discardAll')
-            ->assertSet('pendingActionId', null);
-
-        expect($company->refresh()->status)->toBe(PendingActionStatus::Rejected)
-            ->and($person->refresh()->status)->toBe(PendingActionStatus::Rejected)
-            ->and($task->refresh()->status)->toBe(PendingActionStatus::Rejected)
-            ->and(Company::query()->count())->toBe(0);
-    });
-});
-
-it('refuses a client-set editing target, so one payload cannot open the same field on every step', function (): void {
-    $action = makeBatchCompanyProposal($this->user, ['Alpha', 'Beta']);
-
-    $component = Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $action->getKey(), context: 'conversation');
-
-    // editField() sets both together and clearing resets both; neither is ever
-    // written from the browser. Unlocked, a payload could name a field while
-    // nulling the step, and every step owning that code would render the same
-    // Filament schema against one state path.
-    expect(fn () => $component->set('editingStepId', null))->toThrow(Exception::class)
-        ->and(fn () => $component->set('editingFieldCode', 'name'))->toThrow(Exception::class);
-});
-
-it('ignores skipItem and focusItem against another tenant\'s proposal', function (): void {
-    Bus::fake();
-    $stranger = User::factory()->withPersonalTeam()->create();
-    $foreignRecords = array_map(static fn (string $n): array => ['name' => $n], ['Contoso A', 'Contoso B']);
-    $foreign = proposalCardPa($stranger, ['_batch' => true, 'records' => $foreignRecords], [
-        'title' => 'Create Companies',
-        'summary' => 'Create 2 companies',
-        'items' => [['summary' => 'Contoso A', 'fields' => []], ['summary' => 'Contoso B', 'fields' => []]],
-    ]);
-    $own = makeBatchCompanyProposal($this->user, ['Alpha', 'Beta']);
-
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $own->getKey(), context: 'conversation')
-        ->call('skipItem', (string) $foreign->getKey(), 0)
-        ->assertNotDispatched('proposal:resolved')
-        ->call('focusItem', (string) $foreign->getKey(), 1)
-        ->assertSet('activeStepId', (string) $own->getKey());
-
-    expect($foreign->fresh()->result_data)->toBeNull()
-        ->and($foreign->fresh()->status)->toBe(PendingActionStatus::Pending);
-});
-
-it('never resolves a dock read from a client-supplied id, so another tenant cannot be read through it', function (): void {
-    $stranger = User::factory()->withPersonalTeam()->create();
-    $foreign = proposalCardPa(
-        $stranger,
-        ['name' => 'Contoso Holdings', 'account_owner_id' => (string) $stranger->getKey()],
-        [
-            'title' => 'Create Company',
-            'summary' => 'Contoso Holdings',
-            'fields' => [['label' => 'Name', 'code' => 'name', 'value' => 'Contoso Holdings']],
-        ],
-    );
-
-    $own = proposalCardPa(
-        $this->user,
-        ['name' => 'Acme Corp', 'account_owner_id' => (string) $this->user->getKey()],
-        [
-            'title' => 'Create Company',
-            'summary' => 'Acme Corp',
-            'fields' => [['label' => 'Name', 'code' => 'name', 'value' => 'Acme Corp']],
-        ],
-    );
-
-    // Livewire hands every client-invoked method through implicit route-model
-    // binding, so a public method typed PendingAction would resolve this id with a
-    // bare where(id)->first(). Each read below is called the way the browser calls
-    // it, with the stranger's id in argument position.
-    foreach (['currentRecordFields', 'editableCodes', 'recordCount', 'remainingCount'] as $method) {
-        Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-            ->dispatch('proposal:set-active', id: $own->getKey(), context: 'conversation')
-            ->call($method, $foreign->getKey())
-            ->assertReturned(fn (mixed $returned): bool => ! str_contains(
-                (string) json_encode($returned),
-                'Contoso',
-            ));
-    }
-
-    // The reads answer for the caller's own active step rather than returning
-    // nothing, so the guard cannot be mistaken for the dock simply being broken.
-    Livewire::test(ProposalCard::class, ['context' => 'conversation'])
-        ->dispatch('proposal:set-active', id: $own->getKey(), context: 'conversation')
-        ->call('currentRecordFields', $foreign->getKey())
-        ->assertReturned(fn (array $fields): bool => collect($fields)->contains(
-            fn (array $row): bool => ($row['value'] ?? null) === 'Acme Corp',
-        ));
-});
